feat: moved clal modal out to `ClalModal.tsx`
